More work towards es6/webpack
