+++ src/BloomBrowserUI/bookEdit/EditViewFrame.jade
@@ -9,14 +9,14 @@ html
 		link(rel='stylesheet', href='/bloom/pageChooser/page-chooser.css', type='text/css')
 		link(rel='stylesheet', href='/bloom/bookEdit/bloomDialog.css', type='text/css')
 		script(src='/bloom/node_modules/jquery/dist/jquery.js')
-        script(src='/bloom/bookEdit/js/jquery.hasAttr.js')
-        script(src='/bloom/bookEdit/js/bloomEditing.js')
-        script(src='/bloom/bookEdit/js/interIframeChannel.js')
+		script(src='/bloom/bookEdit/js/jquery.hasAttr.js')
+		script(src='/bloom/bookEdit/js/interIframeChannel.js')
 		
 		//script(src='/bloom/modified_libraries/jquery-ui/jquery-ui-1.10.3.custom.min.js')
 		script(src='/bloom/lib/errorHandler.js')
 		script(src='/bloom/bookEdit/js/editViewFrame.js')
 		//script(src='/bloom/bookEdit/js/calledByCSharp.js')
+		script(src='/bloom/output/commonCode.js')
 		script(src='/bloom/output/pageChooserIFrame.js')
 		script.
 			//var calledByCSharp;
@@ -26,6 +26,11 @@ html
 				// The iframe communication channel needs to be initialized before the child documents are loaded.
 				interIframeChannel = new interIframeChannel();
 				// {simulatedPageFileInBookFolder} is replaced by EditingModel.GetXmlDocumentForEditScreenWebPage().
+				
+				
+				//setting the src's in code here has the effect of waiting until we have the interIFrameChannel ready for them to use
+				//enhance: we could instead bootstrap them here, along with that channel as a parameter, after all three documents are ready
+				// we could do $('iframe').load(function() {.....}) and increment a count and when the count is == number of iframes, then do the bootstraping
 				document.getElementById('page').src = '{simulatedPageFileInBookFolder}';
 				document.getElementById('toolbox').src = '/bloom/toolboxContent';
 			});
