--- src/BloomBrowserUI/bookEdit/js/bloomEditing.js
@@ -1,651 +1,0 @@
-///<reference path="./jquery.hasAttr.d.ts" /> 
-var $ = require('jquery');
-var bloomQtipUtils_1 = require('./bloomQtipUtils');
-var bloomImages_1 = require('./bloomImages');
-var origami_1 = require('./origami');
-var localizationManager_1 = require('../../lib/localizationManager/localizationManager');
-var StyleEditor_1 = require('../StyleEditor/StyleEditor');
-var OverflowChecker_1 = require('../OverflowChecker/OverflowChecker');
-var BloomField_1 = require('../bloomField/BloomField');
-var bloomNotices_1 = require('./bloomNotices');
-var BloomSourceBubbles_1 = require('../sourceBubbles/BloomSourceBubbles');
-var BloomHintBubbles_1 = require('./BloomHintBubbles');
-var getIframeChannel_1 = require('./getIframeChannel');
-var TopicChooser_1 = require('../TopicChooser/TopicChooser');
-/**
- * Fires an event for C# to handle
- * @param {String} eventName
- * @param {String} eventData
- */
-function fireCSharpEditEvent(eventName, eventData) {
-    var event = new MessageEvent(eventName, { 'bubbles': true, 'cancelable': true, 'data': eventData });
-    document.dispatchEvent(event);
-}
-exports.fireCSharpEditEvent = fireCSharpEditEvent;
-function GetDifferenceBetweenHeightAndParentHeight(jqueryNode) {
-    // function also declared and used in StyleEditor
-    if (!jqueryNode) {
-        return 0;
-    }
-    return jqueryNode.parent().height() - jqueryNode.height();
-}
-exports.GetDifferenceBetweenHeightAndParentHeight = GetDifferenceBetweenHeightAndParentHeight;
-function isBrOrWhitespace(node) {
-    return node && ((node.nodeType == 1 && node.nodeName.toLowerCase() == "br") ||
-        (node.nodeType == 3 && /^\s*$/.test(node.nodeValue)));
-}
-function removeTrailingWhiteSpace(node) {
-    if (node && node.nodeType == 3 && node.nodeValue) {
-        // Removes one or more (+) whitespace (\s) at the end ($), across multiple lines (m)
-        node.nodeValue = node.nodeValue.replace(/\s+$/m, "");
-    }
-}
-function TrimTrailingLineBreaksInDivs(node) {
-    //    while ( isBrOrWhitespace(node.firstChild) ) {
-    //        node.removeChild(node.firstChild);
-    //    }
-    while (isBrOrWhitespace(node.lastChild)) {
-        node.removeChild(node.lastChild);
-    }
-    // Without this, FF can display a space which isn't a space due to a trailing \r\n
-    removeTrailingWhiteSpace(node.lastChild);
-}
-function Cleanup() {
-    // for stuff bloom introduces, just use this "bloom-ui" class to have it removed
-    $(".bloom-ui").each(function () {
-        $(this).remove();
-    });
-    bloomQtipUtils_1.default.cleanupBubbles(); // all 3 kinds!
-    $("*.editTimeOnly").remove();
-    $("*.dragHandle").remove();
-    $("*").removeAttr("data-easytabs");
-    $("div.ui-resizable-handle").remove();
-    $('div, figure').each(function () {
-        $(this).removeClass('ui-draggable');
-        $(this).removeClass('ui-resizable');
-        $(this).removeClass('hoverUp');
-    });
-    $('button').each(function () {
-        $(this).remove();
-    });
-    $('div.bloom-editable').each(function () {
-        TrimTrailingLineBreaksInDivs(this);
-    });
-    bloomImages_1.cleanupImages();
-    origami_1.cleanupOrigami();
-}
-function GetStyleClassFromElement(element) {
-    var c = $(element).attr("class");
-    if (!c)
-        c = "";
-    var classes = c.split(' ');
-    for (var i = 0; i < classes.length; i++) {
-        if (classes[i].indexOf('-style') > 0) {
-            return classes[i];
-        }
-    }
-    return null;
-}
-//add a delete button which shows up when you hover
-function SetupDeletable(containerDiv) {
-    $(containerDiv).mouseenter(function () {
-        var button = $("<button class='deleteButton smallImageButton' title='Delete'></button>");
-        $(button).click(function () {
-            $(containerDiv).remove();
-        });
-        $(this).prepend(button);
-    })
-        .mouseleave(function () {
-        $(this).find(".deleteButton").each(function () {
-            $(this).remove();
-        });
-    });
-    return $(containerDiv);
-}
-// Add various editing key handlers
-function AddEditKeyHandlers(container) {
-    //Make F6 apply a superscript style (later we'll change to ctrl+shift+plus, as word does. But capturing those in js by hand is a pain.
-    //nb: we're avoiding ctrl+plus and ctrl+shift+plus (as used by MS Word), because they means zoom in browser. also three keys is too much
-    $(container).find("div.bloom-editable").on('keydown', null, 'F6', function (e) {
-        var selection = document.getSelection();
-        if (selection) {
-            //NB: by using exeCommand, we get undo-ability
-            document.execCommand("insertHTML", false, "<span class='superscript'>" + document.getSelection() + "</span>");
-        }
-    });
-    $(container).find("div.bloom-editable").on('keydown', null, 'ALT+CTRL+0', function (e) {
-        e.preventDefault();
-        document.execCommand("formatBlock", false, "P");
-    });
-    // Make F7 apply top-level header style (H1)
-    $(container).find("div.bloom-editable").on('keydown', null, 'F7', function (e) {
-        e.preventDefault();
-        document.execCommand("formatBlock", false, "H1");
-    });
-    $(container).find("div.bloom-editable").on('keydown', null, 'ALT+CTRL+1', function (e) {
-        e.preventDefault();
-        document.execCommand("formatBlock", false, "H1");
-    });
-    // Make F8 apply header style (H2)
-    $(container).find("div.bloom-editable").on('keydown', null, 'F8', function (e) {
-        e.preventDefault();
-        document.execCommand("formatBlock", false, "H2");
-    });
-    $(container).find("div.bloom-editable").on('keydown', null, 'ALT+CTRL+2', function (e) {
-        e.preventDefault();
-        document.execCommand("formatBlock", false, "H2");
-    });
-    $(document).bind('keydown', 'ctrl+space', function (e) {
-        e.preventDefault();
-        document.execCommand("removeFormat", false, false); //will remove bold, italics, etc. but not things that use elements, like h1
-    });
-    $(document).bind('keydown', 'ctrl+u', function (e) {
-        e.preventDefault();
-        document.execCommand("underline", null, null);
-    });
-    $(document).bind('keydown', 'ctrl+b', function (e) {
-        e.preventDefault();
-        document.execCommand("bold", null, null);
-    });
-    $(document).bind('keydown', 'ctrl+i', function (e) {
-        e.preventDefault();
-        document.execCommand("italic", null, null);
-    });
-    //note: these have the effect of introducing a <div> inside of the div.bloom-editable we're in.
-    $(document).bind('keydown', 'ctrl+r', function (e) {
-        e.preventDefault();
-        document.execCommand("justifyright", false, null);
-    });
-    $(document).bind('keydown', 'ctrl+l', function (e) {
-        e.preventDefault();
-        document.execCommand("justifyleft", false, null);
-    });
-    $(document).bind('keydown', 'ctrl+shift+e', function (e) {
-        e.preventDefault();
-        document.execCommand("justifycenter", false, null);
-    });
-    //Note, CTRL+N is also caught, but up on the Shell where it is turned into an event, so that it can be caught even when the focus isn't on the browser
-}
-// Add little language tags
-function AddLanguageTags(container) {
-    $(container).find(".bloom-editable:visible[contentEditable=true]").each(function () {
-        var $this = $(this);
-        // If this DIV already had a language tag, remove the content in case we decide the situation has changed.
-        if ($this.hasAttr('data-languageTipContent')) {
-            $this.removeAttr('data-languageTipContent');
-        }
-        // With a really small box that also had a hint qtip, there wasn't enough room and the two fought
-        // with each other, leading to flashing back and forth
-        // Of course that was from when Language Tags were qtips too, but I think I'll leave the restriction for now.
-        if ($this.width() < 100) {
-            return;
-        }
-        // Make sure language tags appear or disappear depending on what edit mode we are in
-        var isTranslationMode = IsInTranslationMode();
-        if (isTranslationMode && $this.hasClass('bloom-readOnlyInTranslationMode')) {
-            return;
-        }
-        if (!isTranslationMode && $this.hasClass('bloom-readOnlyInAuthorMode')) {
-            return;
-        }
-        var key = $this.attr("lang");
-        if (key == "*" || key.length < 1)
-            return; //seeing a "*" was confusing even to me
-        // if this or any parent element has the class bloom-hideLanguageNameDisplay, we don't want to show any of these tags
-        // first usage (for instance) was turning off language tags for a whole page
-        if ($this.hasClass('bloom-hideLanguageNameDisplay') || $this.parents('.bloom-hideLanguageNameDisplay').length != 0) {
-            return;
-        }
-        var whatToSay = localizationManager_1.default.getText(key);
-        if (!whatToSay)
-            whatToSay = key; //just show the code
-        // Put whatToSay into data attribute for pickup by the css
-        $this.attr('data-languageTipContent', whatToSay);
-    });
-}
-function SetBookCopyrightAndLicenseButtonVisibility(container) {
-    var shouldShowButton = !($(container).find("DIV.copyright").text());
-    $(container).find("button#editCopyrightAndLicense").css("display", shouldShowButton ? "inline" : "none");
-}
-function DecodeHtml(encodedString) {
-    return encodedString.replace(/&amp;/g, '&').replace(/&lt;/g, '<').replace(/&gt;/g, '>').replace(/&#39;/g, "'").replace(/&#169;/g, "©");
-}
-function GetEditor() {
-    return new StyleEditor_1.default("/bloom/bookEdit");
-}
-function GetOverflowChecker() {
-    return new OverflowChecker_1.default();
-}
-function IsInTranslationMode() {
-    var body = $("body");
-    if (!body.hasAttr('editMode'))
-        return false;
-    else {
-        return body.attr('editMode') == "translation";
-    }
-}
-function getToolbox() {
-    // I (GJM) tried to define this is readerTools.ts, but it wasn't loaded if no reader tools were active!
-    var toolbox = parent.window.document.getElementById("toolbox");
-    // toolbox will be undefined during unit testing
-    if (toolbox)
-        return toolbox.contentWindow['toolbox'];
-}
-// Originally, all this code was in document.load and the selectors were acting
-// on all elements (not bound by the container).  I added the container bound so we
-// can add new elements (such as during layout mode) and call this on only newly added elements.
-// Now document.load calls this with $('body') as the container.
-// REVIEW: Some of these would be better off in OneTimeSetup, but too much risk to try to decide right now.
-function SetupElements(container) {
-    bloomImages_1.SetupImagesInContainer(container);
-    //add a marginBox if it's missing. We introduced it early in the first beta
-    $(container).find(".bloom-page").each(function () {
-        if ($(this).find(".marginBox").length == 0) {
-            $(this).wrapInner("<div class='marginBox'></div>");
-        }
-    });
-    $(container).find(".bloom-editable").each(function () { BloomField_1.default.ManageField(this); });
-    //make textarea edits go back into the dom (they were designed to be POST'ed via forms)
-    $(container).find("textarea").blur(function () {
-        this.innerHTML = this.value;
-    });
-    var toolbox = getToolbox();
-    var toolboxVisible = false;
-    // toolbox might be undefined in unit testing?
-    //TODO: how to properly access toolbox from here? this is yet another *#$!% global in our code, grrrr
-    if (toolbox) {
-    }
-    SetBookCopyrightAndLicenseButtonVisibility(container);
-    //CSS normally can't get at the text in order to, for example, show something different if it is empty.
-    //This allows you to add .bloom-needs-data-text to a bloom-translationGroup in order to get
-    //its child bloom-editable's to have data-texts's on them
-    $(container).find(".bloom-translationGroup.bloom-text-for-css .bloom-editable").each(function () {
-        // initially fill it
-        $(this).attr('data-text', this.textContent);
-        // keep it up to date
-        $(this).on('blur paste input', function () {
-            $(this).attr('data-text', this.textContent);
-        });
-    });
-    //in bilingual/trilingual situation, re-order the boxes to match the content languages, so that stylesheets don't have to
-    $(container).find(".bloom-translationGroup").each(function () {
-        var contentElements = $(this).find("textarea, div.bloom-editable");
-        contentElements.sort(function (a, b) {
-            //using negatives so that something with none of these labels ends up with a > score and at the end
-            //reviewSlog
-            var scoreA = ($(a).hasClass('bloom-content1') ? 1 : 0) * -3 + ($(a).hasClass('bloom-content2') ? 1 : 0 * -2) + ($(a).hasClass('bloom-content3') ? 1 : 0 * -1);
-            var scoreB = ($(b).hasClass('bloom-content1') ? 1 : 0) * -3 + ($(b).hasClass('bloom-content2') ? 1 : 0 * -2) + ($(b).hasClass('bloom-content3') ? 1 : 0 * -1);
-            if (scoreA < scoreB)
-                return -1;
-            if (scoreA > scoreB)
-                return 1;
-            return 0;
-        });
-        //do the actual rearrangement
-        $(this).append(contentElements);
-    });
-    //Convert Standard Format Markers in the pasted text to html spans
-    $(container).find("div.bloom-editable").on("paste", function (e) {
-        var theEvent = e.originalEvent;
-        if (!theEvent.clipboardData)
-            return;
-        var s = theEvent.clipboardData.getData('text/plain');
-        if (s == null || s == '')
-            return;
-        var re = new RegExp('\\\\v\\s(\\d+)', 'g');
-        var matches = re.exec(s);
-        if (matches == null) {
-        }
-        else {
-            e.preventDefault();
-            var x = s.replace(re, "<span class='superscript'>$1</span>");
-            document.execCommand("insertHtml", false, x);
-        }
-    });
-    AddEditKeyHandlers(container);
-    //--------------------------------
-    //keep divs vertically centered (yes, I first tried *all* the css approaches, they don't work for our situation)
-    //do it initially
-    $(container).find(".bloom-centerVertically").CenterVerticallyInParent();
-    //reposition as needed
-    $(container).find(".bloom-centerVertically").resize(function () {
-        $(this).CenterVerticallyInParent();
-    });
-    //html5 provides for a placeholder attribute, but not for contenteditable divs like we use.
-    //So one of our foundational stylesheets looks for @data-placeholder and simulates the
-    //@placeholder behavior.
-    //Now, what's going on here is that we also support
-    //<label class='placeholder'> inside a div.bloom-translationGroup to get this placeholder
-    //behavior on each of the fields inside the group .
-    //Using <label> instead of the attribute makes the html much easier to read, write, and add additional
-    //behaviors through classes.
-    //So the job of this bit here is to take the label.placeholder and create the data-placeholders.
-    $(container).find("*.bloom-translationGroup > label.placeholder").each(function () {
-        var labelText = $(this).text();
-        //put the attributes on the individual child divs
-        $(this).parent().find('.bloom-editable').each(function () {
-            //enhance: it would make sense to allow each of these to be customized for their div
-            //so that you could have a placeholder that said "Name in {lang}", for example.
-            $(this).attr('data-placeholder', labelText);
-            //next, it's up to CSS to draw the placeholder when the field is empty.
-        });
-    });
-    $(container).find('div.bloom-editable').each(function () {
-        $(this).attr('contentEditable', 'true');
-    });
-    // Bloom needs to make some fields readonly. E.g., the original license when the user is translating a shellbook
-    // Normally, we'd control this is a style in editTranslationMode.css/editOriginalMode.css. However, "readonly" isn't a style, just
-    // an attribute, so it can't be included in css.
-    // The solution here is to add the readonly attribute when we detect that the css has set the cursor to "not-allowed".
-    $(container).find('textarea, div').focus(function () {
-        //        if ($(this).css('border-bottom-color') == 'transparent') {
-        if ($(this).css('cursor') == 'not-allowed') {
-            $(this).attr("readonly", "true");
-            $(this).removeAttr("contentEditable");
-        }
-        else {
-            $(this).removeAttr("readonly");
-        }
-    });
-    AddLanguageTags(container);
-    // If the user moves over something they can't edit, show a tooltip explaining why not
-    bloomNotices_1.default.addEditingNotAllowedMessages(container);
-    //Same thing for divs which are potentially editable, but via the contentEditable attribute instead of TextArea's ReadOnly attribute
-    // editTranslationMode.css/editOriginalMode.css can't get at the contentEditable (css can't do that), so
-    // so they set the cursor to "not-allowed", and we detect that and set the contentEditable appropriately
-    $(container).find('div.bloom-readOnlyInTranslationMode').focus(function () {
-        if ($(this).css('cursor') == 'not-allowed') {
-            $(this).removeAttr("contentEditable");
-        }
-        else {
-            $(this).attr("contentEditable", "true");
-        }
-    });
-    //first used in the Uganda SHRP Primer 1 template, on the image on day 1
-    $(container).find(".bloom-draggableLabel").each(function () {
-        // previous to June 2014, containment was not working, so some items may be
-        // out of bounds. Or the stylesheet could change the size of things. This gets any such back in bounds.
-        if ($(this).position().left < 0) {
-            $(this).css('left', 0);
-        }
-        if ($(this).position().top < 0) {
-            $(this).css('top', 0);
-        }
-        if ($(this).position().left + $(this).width() > $(this).parent().width()) {
-            $(this).css('left', $(this).parent().width() - $(this).width());
-        }
-        if ($(this).position().top > $(this).parent().height()) {
-            $(this).css('top', $(this).parent().height() - $(this).height());
-        }
-        $(this).draggable({
-            containment: "parent",
-            handle: '.dragHandle'
-        });
-    });
-    $(container).find(".bloom-draggableLabel")
-        .mouseenter(function () {
-        $(this).prepend(" <div class='dragHandle'></div>");
-    });
-    $(container).find(".bloom-draggableLabel").mouseleave(function () {
-        $(this).find(".dragHandle").each(function () {
-            $(this).remove();
-        });
-    });
-    bloomQtipUtils_1.default.repositionPictureDictionaryTooltips(container);
-    /* Support in page combo boxes that set a class on the parent, thus making some change in the layout of the pge.
-    Example:
-         <select name="Story Style" class="bloom-classSwitchingCombobox">
-             <option value="Fictional">Fiction</option>
-             <option value="Informative">Informative</option>
-     </select>
-     */
-    //First we select the initial value based on what class is currently set, or leave to the default if none of them
-    $(container).find(".bloom-classSwitchingCombobox").each(function () {
-        //look through the classes of the parent for any that match one of our combobox values
-        var i;
-        for (i = 0; i < this.options.length; i++) {
-            var c = this.options[i].value;
-            if ($(this).parent().hasClass(c)) {
-                $(this).val(c);
-                break;
-            }
-        }
-    });
-    //And now we react to the user choosing a different value
-    $(container).find(".bloom-classSwitchingCombobox").change(function () {
-        //remove any of the values that might already be set
-        var i;
-        for (i = 0; i < this.options.length; i++) {
-            var c = this.options[i].value;
-            $(this).parent().removeClass(c);
-        }
-        //add back in the one they just chose
-        $(this).parent().addClass(this.value);
-    });
-    //only make things deletable if they have the deletable class *and* page customization is enabled
-    $(container).find("DIV.bloom-page.bloom-enablePageCustomization DIV.bloom-deletable").each(function () {
-        SetupDeletable(this);
-    });
-    bloomNotices_1.default.addExperimentalNotice(container); // adds notice to Picture Dictionary pages
-    $(container).find(".bloom-resizable").each(function () {
-        bloomImages_1.SetupResizableElement(this);
-    });
-    bloomImages_1.SetOverlayForImagesWithoutMetadata(container);
-    //note, the normal way is for the user to click the link on the bubble.
-    //But clicking on the existing topic may be natural too, and this prevents
-    //them from editing it by hand.
-    $(container).find("div[data-book='topic']").click(function () {
-        if ($(this).css('cursor') == 'not-allowed')
-            return;
-        TopicChooser_1.default.showTopicChooser();
-    });
-    // Copy source texts out to their own div, where we can make a bubble with tabs out of them
-    // We do this because if we made a bubble out of the div, that would suck up the vernacular editable area, too,
-    if ($(container).find(".bloom-preventSourceBubbles").length == 0) {
-        $(container).find("*.bloom-translationGroup").not(".bloom-readOnlyInTranslationMode").each(function () {
-            if ($(this).find("textarea, div").length > 1) {
-                BloomSourceBubbles_1.default.ProduceSourceBubbles(this);
-            }
-        });
-    }
-    //NB: this should be after the ProduceSourceBubbles(), so that it can remove labels that
-    //would otherwise be underfoot when we call this. This would happen with the Book Title
-    //when there are source languages to show
-    BloomHintBubbles_1.default.addHintBubbles(container);
-    // Add overflow event handlers so that when a div is overfull,
-    // we add the overflow class and it gets a red background or something
-    // Moved AddOverflowHandlers() after SetupImage() because some pages with lots of placeholders
-    // were prematurely overflowing before the images were set to the right size.
-    GetOverflowChecker().AddOverflowHandlers(container);
-    var editor = GetEditor();
-    $(container).find("div.bloom-editable:visible").each(function () {
-        // If the .bloom-editable or any of its ancestors (including <body>) has the class "bloom-userCannotModifyStyles",
-        // then the controls that allow the user to adjust the styles will not be shown.This does not prevent the user
-        // from doing character styling, e.g. CTRL+b for bold.
-        if ($(this).closest('.bloom-userCannotModifyStyles').length == 0) {
-            $(this).focus(function () {
-                editor.AttachToBox(this);
-            });
-        }
-    });
-    loadLongpressInstructions($(container).find('.bloom-editable'));
-    //When we do a CTRL+A DEL, FF leaves us with a <br></br> at the start. When the first key is then pressed,
-    //a blank line is shown and the letter pressed shows up after that.
-    //This detects that situation when we type the first key after the deletion, and first deletes the <br></br>.
-    $(container).find('.bloom-editable').keypress(function (event) {
-        //this is causing a worse problem, (preventing us from typing empty lines to move the start of the text down), so we're going to live with the empty space for now.
-        // TODO: perhaps we can act when the DEL or Backspace occurs and then detect this situation and clean it up.
-        //         if ($(event.target).text() == "") { //NB: the browser inspector shows <br></br>, but innerHTML just says "<br>"
-        //            event.target.innerHTML = "";
-        //        }
-    });
-    //This detects that situation when we do CTRL+A and then type a letter, instead of DEL
-    $(container).find('.bloom-editable').keyup(function (event) {
-        //console.log(event.target.innerHTML);
-        // If they pressed a letter instead of DEL, we get this case:
-        if ($(event.target).find("#formatButton").length == 0) {
-            //they have also deleted the formatButton, so put it back in
-            // console.log('attaching'); REVIEW: this shows that we're doing the attaching on the first character entered, even though it appears the editor was already attached.
-            //so we actually attach twice. That's ok, the editor handles that, but I don't know why we're passing the if, and it could be improved.
-            if ($(this).closest('.bloom-userCannotModifyStyles').length == 0)
-                editor.AttachToBox(this);
-        }
-        else {
-            // already have a format cog, better make sure it's in the right place
-            editor.AdjustFormatButton($(this));
-        }
-    });
-    // focus on the first editable field
-    // HACK for BL-1139: except for some reason when the Reader tools are active this causes
-    // quick typing on a newly loaded page to get the cursor messed up. So for the Reader tools, the
-    // user will need to actually click in the div to start typing.
-    if (!toolboxVisible)
-        $(container).find("textarea, div.bloom-editable").first().focus(); //review: this might choose a textarea which appears after the div. Could we sort on the tab order?
-}
-// Only put setup code here which is guaranteed to only be run once per page load.
-// e.g. Don't put setup for elements such as image containers or editable boxes which may get added after page load.
-function OneTimeSetup() {
-    origami_1.setupOrigami();
-}
-// ---------------------------------------------------------------------------------
-// document ready function
-// ---------------------------------------------------------------------------------
-$(document).ready(function () {
-    bloomQtipUtils_1.default.setQtipZindex();
-    $.fn.reverse = function () {
-        return this.pushStack(this.get().reverse(), arguments);
-    };
-    // Attach a function to implement zooming on mouse wheel with ctrl
-    $('body').on('wheel', function (e) {
-        var theEvent = e.originalEvent;
-        // reviewSlog does this application work?
-        if (!theEvent.ctrlKey)
-            return;
-        // look for an existing transform:scale setting and extract the scale. If not found, use 1.0 as starting point.
-        var styleString = $('body').attr('style');
-        var scale = 1.0;
-        var searchData = /scale\(([^,]*),/.exec(styleString);
-        if (searchData) {
-            scale = parseFloat(searchData[1]);
-        }
-        // Dividing by 20 seems to make it zoom at a manageable rate, at least with my mouse.
-        // The limitation to zooming between 1/3 and 3 times is arbitrary.
-        scale = Math.min(Math.max(scale - theEvent.deltaY / 20, 0.33), 3.0);
-        // This works with a rule in editMode.less that makes the transform's origin the top left
-        $('body').attr('style', 'transform: scale(' + scale + ',' + scale + ')');
-    });
-    /* reviewSlog typescript just couldn't cope with this. Our browser has this built in , so it's ok
-        //if this browser doesn't have endsWith built in, add it
-        if (typeof String.prototype.endsWith !== 'function') {if (typeof String.prototype.endsWith !== 'function') {
-            String.prototype.endsWith = function (suffix) {
-                return this.indexOf(suffix, this.length - suffix.length) !== -1;
-            };
-        }
-    
-        // Defines a starts-with function
-        if (typeof String.prototype.startsWith != 'function') {
-            String.prototype.startsWith = function (str) {
-                return this.indexOf(str) == 0;
-            };
-        }
-    */
-    //eventually we want to run this *after* we've used the page, but for now, it is useful to clean up stuff from last time
-    Cleanup();
-    SetupElements($('body'));
-    OneTimeSetup();
-    // configure ckeditor
-    if (typeof CKEDITOR === "undefined")
-        return; // this happens during unit testing
-    CKEDITOR.disableAutoInline = true;
-    // Map from ckeditor id strings to the div the ckeditor is wrapping.
-    var mapCkeditDiv = new Object();
-    // attach ckeditor to the contenteditable="true" class="bloom-content1"
-    // also to contenteditable="true" and class="bloom-content2" or class="bloom-content3"
-    // but skip any element with class="bloom-userCannotModifyStyles"
-    $('div.bloom-page').find('.bloom-content1[contenteditable="true"],.bloom-content2[contenteditable="true"],.bloom-content3[contenteditable="true"],.bloom-contentNational1[contenteditable="true"]').each(function () {
-        if ($(this).hasClass('bloom-userCannotModifyStyles'))
-            return; // equivalent to 'continue'
-        if ($(this).css('cursor') == 'not-allowed')
-            return;
-        var ckedit = CKEDITOR.inline(this);
-        // Record the div of the edit box for use later in positioning the format bar.
-        mapCkeditDiv[ckedit.id] = this;
-        // show or hide the toolbar when the text selection changes
-        ckedit.on('selectionCheck', function (evt) {
-            var editor = evt['editor'];
-            // Length of selected text is more reliable than comparing
-            // endpoints of the first range.  Mozilla can return multiple
-            // ranges with the first one being empty.
-            var selection = editor.getSelection();
-            var textSelected = selection.getSelectedText();
-            var show = (textSelected.length > 0);
-            var bar = $('body').find('.' + editor.id);
-            show ? bar.show() : bar.hide();
-            // Move the format bar on the screen if needed.
-            // (Note that offsets are not defined if it's not visible.)
-            if (show) {
-                var barTop = bar.offset().top;
-                var div = mapCkeditDiv[editor.id];
-                var rect = div.getBoundingClientRect();
-                var parent = bar.scrollParent();
-                var scrollTop = (parent) ? parent.scrollTop() : 0;
-                var boxTop = rect.top + scrollTop;
-                if (boxTop - barTop < 5) {
-                    var barLeft = bar.offset().left;
-                    var barHeight = bar.height();
-                    bar.offset({ top: boxTop - barHeight, left: barLeft });
-                }
-            }
-        });
-        // hide the toolbar when ckeditor starts
-        ckedit.on('instanceReady', function (evt) {
-            var editor = evt['editor'];
-            var bar = $('body').find('.' + editor.id);
-            bar.hide();
-        });
-        BloomField_1.default.WireToCKEditor(this, ckedit);
-    });
-    //this is some sample code for working on CommandAvailabilityPublisher websocket messages
-    //   var client = new WebSocket("ws://127.0.0.1:8189");
-    //   client.onmessage = function(event) {
-    //        var commandStatus = JSON.parse(event.data);
-    //        alert("DeleteCurrentPage Command "+ (commandStatus.deleteCurrentPage.enabled == true ? "Enabled" : "Disabled")) ;
-    //    }
-}); // end document ready function
-// This is invoked from C# when we are about to change pages. It is mainly for origami,
-// but preparePageForEditingAfterOrigamiChangesEvent currently has the (very important)
-// side effect of saving the changes to the current page.
-var pageSelectionChanging = function () {
-    var marginBox = $('.marginBox');
-    marginBox.removeClass('origami-layout-mode');
-    marginBox.find('.bloom-translationGroup .textBox-identifier').remove();
-    fireCSharpEditEvent('finishSavingPage', '');
-};
-// This is invoked from C# when we are about to leave a page (often right after the previous
-// method for changing pages).  It is mainly to clean things up so that garbage collection
-// won't lose multiple megabytes of data that both the DOM (C++) and Javascript subsystems
-// think the other is still using.
-var disconnectForGarbageCollection = function () {
-    // disconnect all event handlers
-    //review: was this, but TS didn't like it    $.find().off();
-    $('body').find('*').off();
-    var page = $('.bloom-page');
-    // blow away any img elements to ensure their data disappears.
-    // (the whole document is being replaced, and this happens after it's been saved to a file)
-    page.find('img').each(function () { $(this).attr("src", ""); });
-    page.find('img').each(function () { $(this).remove(); });
-    $('[style*=".background-image"]').each(function () { $(this).remove(); });
-};
-function loadLongpressInstructions(jQuerySetOfMatchedElements) {
-    getIframeChannel_1.default().simpleAjaxGet('/bloom/windows/useLongpress', function (response) {
-        if (response === 'Yes') {
-            localizationManager_1.default.asyncGetText('BookEditor.CharacterMap.Instructions', 'To select, use your mouse wheel or point at what you want, then release the key.')
-                .done(function (translation) {
-                jQuerySetOfMatchedElements.longPress({ instructions: "<div class='instructions'>" + translation + "</div>" });
-            });
-        }
-        ;
-    });
-}
-
-//# sourceMappingURL=bloomEditing.js.map
