+++ src/BloomBrowserUI/bookEdit/toolbox/toolbox.jade
@@ -3,24 +3,25 @@ html
 		meta(charset='UTF-8')
 		title Content of the Toolbox view (far right pane in book edit)
 		link(rel='stylesheet', href='/bloom/themes/bloom-jqueryui-theme/jquery-ui-1.8.16.custom.css')
-		script(src='/bloom/output/commonCode.js')
-		script(src='/bloom/output/toolboxIFrame.js')
-
-		//TODO: we should not be need to explicitly import this here
-		script(src="/bloom/node_modules/jquery/dist/jquery.min.js")
-
 		link(rel='stylesheet', href='/bloom/bookEdit/html/font-awesome/css/font-awesome.min.css')
 		link(rel='stylesheet', href='/bloom/bookEdit/toolbox/toolbox.css')
 		link(rel='stylesheet', href='/bloom/bookEdit/toolbox/readerTools.css')
 		link(rel='stylesheet', href='/bloom/bookEdit/toolbox/talkingBook/audioRecording.css')
+		link(rel='stylesheet', href='/bloom/lib/jquery.qtip.css')
+
+		script(src='/bloom/output/commonCode.js')
+		script(src='/bloom/output/toolboxIFrame.js')
+
+		//TODO: we should not be need to explicitly import this here
+		//script(src="/bloom/node_modules/jquery/dist/jquery.min.js")
+
 		
 		//TODO: get these raw script calls out of here, should be part of toolboxIFrame.js
 
 		// not sure we need all three qtip files are needed. qtips are only in the toolbox so far.
 		// Something is needed so that $(...).qtip works when manipulating the main page DOM.
-		script(src='/bloom/lib/jquery.qtip.js')
-		script(src='/bloom/lib/jquery.qtipSecondary.js')
-		link(rel='stylesheet', href='/bloom/lib/jquery.qtip.css')
+		// script(src='/bloom/lib/jquery.qtip.js')
+		// script(src='/bloom/lib/jquery.qtipSecondary.js')
 	body
 		.toolboxRoot
 			#toolbox
