+++ src/BloomBrowserUI/bookEdit/toolbox/toolboxBootstrap.ts
@@ -1,0 +1,10 @@
+/// <reference path="../../typings/jquery/jquery.d.ts" />
+import * as $ from 'jquery';
+import {restoreToolboxSettings} from './toolbox';
+
+//this is currently inserted by c#. TODO: get settings via ajax
+declare function GetToolboxSettings():any;
+
+$(document).ready(function() { 
+    restoreToolboxSettings(GetToolboxSettings()); 
+});
